+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Fixed
+
+- `Beaker.workspace.clear()` will remove uncommitted datasets too.
+
 ## [v1.16.0](https://github.com/allenai/beaker-py/releases/tag/v1.16.0) - 2023-01-26
 
 ### Added